+++ Google_tests/Tests_Dual_Graph.cpp
@@ -1820,14 +1820,14 @@ TEST_F(MGridGen_Dual_Graph, compute_neighbourhood_of_cc_MGridGen_case_0) {
     unsigned short goal_card = 8;
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated);
 
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1,  1},
                                                       {2,  1},
                                                       {3,  2},
                                                       {4,  3},
@@ -1848,13 +1848,13 @@ TEST_F(MGridGen_Dual_Graph, compute_neighbourhood_of_cc_MGridGen_case_1) {
     unsigned short goal_card = 8;
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0, 1};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{2,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{2,  1},
                                                       {3,  1},
                                                       {4,  3},
                                                       {5,  2},
@@ -1874,13 +1874,13 @@ TEST_F(MGridGen_Dual_Graph, compute_neighbourhood_of_cc_MGridGen_case_2) {
     unsigned short goal_card = 8;
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0, 1, 2, 3, 6};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{4,  2},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{4,  2},
                                                       {5,  1},
                                                       {7,  3},
                                                       {8,  2},
@@ -1901,14 +1901,14 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_case_3_cc_not_connect
     unsigned short goal_card = 8;
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0, 1, 11};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated
     );
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{2,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{2,  1},
                                                       {3,  1},
                                                       {4,  3},
                                                       {5,  2},
@@ -1932,13 +1932,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints) {
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 4, 7, 14, 2, 0, 12, 13, 5, 6, 1, 11, 3, 8, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1,  1},
                                                       {2,  1},
                                                       {3,  2},
                                                       {4,  3},
@@ -1960,13 +1960,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints_11) {
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 4, 7, 2, 0, 12, 13, 5, 6, 1, 3, 8, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1, 1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1, 1},
                                                       {2, 1},
                                                       {3, 2},
                                                       {4, 3},
@@ -1987,13 +1987,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints_118) {
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 4, 7, 2, 0, 12, 13, 5, 6, 1, 3, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1, 1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1, 1},
                                                       {2, 1},
                                                       {3, 2},
                                                       {4, 3},
@@ -2014,13 +2014,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints_118_max
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 4, 7, 2, 0, 12, 13, 5, 6, 1, 3, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1,  1},
                                                       {2,  1},
                                                       {3,  2},
                                                       {4,  3},
@@ -2044,13 +2044,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints_1184) {
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 7, 2, 0, 12, 13, 5, 6, 1, 3, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{1, 1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{1, 1},
                                                       {2, 1},
                                                       {3, 2},
                                                       {5, 2},
@@ -2069,13 +2069,13 @@ TEST_F(MGridGen_Dual_Graph, compute_of_neighbourhood_of_cc_w_constraints_13) {
     unsigned short number_of_order_of_neighbourhood = 3;
     unordered_set<long> s_seeds = {0};
     unordered_set<long> s_of_constrained_fc = {9, 4, 7, 14, 2, 0, 12, 13, 5, 6, 11, 8, 10};
-    unordered_map<long, int> d_neighbours_of_seed = {};
+    unordered_map<long, unsigned short> d_neighbours_of_seed = {};
     (*g).compute_neighbourhood_of_cc(s_seeds,
                                      number_of_order_of_neighbourhood,
                                      d_neighbours_of_seed,
                                      goal_card,
                                      a_is_fc_agglomerated, s_of_constrained_fc);
-    unordered_map<long, int> ref_dict_fc_n_of_seed = {{2,  1},
+    unordered_map<long, unsigned short> ref_dict_fc_n_of_seed = {{2,  1},
                                                       {4,  3},
                                                       {5,  2},
                                                       {6,  2},
